+++ tests/server_event.rs
@@ -245,6 +245,51 @@ fn event_queue() {
 }
 
 #[test]
+fn independent_event() {
+    let mut server_app = App::new();
+    let mut client_app = App::new();
+    for app in [&mut server_app, &mut client_app] {
+        app.add_plugins((
+            MinimalPlugins,
+            RepliconPlugins.set(ServerPlugin {
+                tick_policy: TickPolicy::EveryFrame,
+                ..Default::default()
+            }),
+        ))
+        .replicate::<DummyComponent>()
+        .add_server_event::<DummyEvent>(ChannelKind::Ordered)
+        .make_independent::<DummyEvent>();
+    }
+
+    server_app.connect_client(&mut client_app);
+
+    // Spawn entity to trigger world change.
+    server_app.world_mut().spawn((Replicated, DummyComponent));
+
+    server_app.update();
+    server_app.exchange_with_client(&mut client_app);
+    client_app.update();
+    server_app.exchange_with_client(&mut client_app);
+
+    // Artificially reset the init tick.
+    // Normal events would be queued and not triggered yet,
+    // but our independent event should be triggered immediately.
+    *client_app.world_mut().resource_mut::<ServerInitTick>() = Default::default();
+    server_app.world_mut().send_event(ToClients {
+        mode: SendMode::Broadcast,
+        event: DummyEvent,
+    });
+
+    server_app.update();
+    server_app.exchange_with_client(&mut client_app);
+    client_app.update();
+
+    // Event should have already been triggered, even without resetting the tick,
+    // since it's independent.
+    assert_eq!(client_app.world().resource::<Events<DummyEvent>>().len(), 1);
+}
+
+#[test]
 fn different_ticks() {
     let mut server_app = App::new();
     let mut client_app1 = App::new();