+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `ServerEventAppExt::make_independent` to let events be triggered without waiting for replication on the same tick.
+
 ## [0.27.0] - 2024-07-04
 
 ### Changed